third_party/WebKit/LayoutTests/fast/dom/NamedNodeMap-missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
[
    "setNamedItem()",
    "setNamedItemNS()",
].forEach(function(expr)
{
    window.attributes = document.body.attributes;
    shouldThrow("attributes." + expr);
});
</script>
</body>
</html>
